+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/UI/UIImage/UIImage.cs
@@ -71,7 +71,6 @@ namespace DCL.Components
                     });
 
                     fetchRoutine = CoroutineStarter.Start(fetchIEnum);
-                    return null;
                 }
             }
             else
